+++ circuitbreaker/openstate_test.go
@@ -27,3 +27,35 @@ func TestTryAcquirePermit(t *testing.T) {
 	assert.True(t, breaker.TryAcquirePermit())
 	assert.True(t, breaker.IsHalfOpen())
 }
+
+func TestRemainingDelay(t *testing.T) {
+	breaker := Builder[any]().WithDelayFunc(func(exec failsafe.ExecutionAttempt[any]) time.Duration {
+		return 1 * time.Second
+	}).Build().(*circuitBreaker[any])
+	breaker.open(testutil.TestExecution[any]{})
+
+	// When / Then
+	remainingDelay := breaker.RemainingDelay()
+	assert.True(t, remainingDelay > 0)
+	assert.True(t, remainingDelay.Milliseconds() < 1001)
+
+	time.Sleep(110 * time.Millisecond)
+	remainingDelay = breaker.RemainingDelay()
+	assert.True(t, remainingDelay > 0)
+	assert.True(t, remainingDelay.Milliseconds() < 900)
+}
+
+func TestNoRemainingDelay(t *testing.T) {
+	breaker := Builder[any]().WithDelayFunc(func(exec failsafe.ExecutionAttempt[any]) time.Duration {
+		return 10 * time.Millisecond
+	}).Build().(*circuitBreaker[any])
+	assert.Equal(t, time.Duration(0), breaker.RemainingDelay())
+
+	// When
+	breaker.open(testutil.TestExecution[any]{})
+	assert.True(t, breaker.RemainingDelay() > 0)
+	time.Sleep(50 * time.Millisecond)
+
+	// Then
+	assert.Equal(t, time.Duration(0), breaker.RemainingDelay())
+}